Extract a "main" function which sets up signal interrupts
See https://github.com/thefrontside/effection.js/issues/68
